show wallet details

+++ shared/domain/src/commonMain/kotlin/com/itcluster/mobile/presentation/models/WalletTransactionModel.kt
@@ -3,13 +3,13 @@ package com.itcluster.mobile.presentation.models
 class WalletTransactionModel(
     val transactionId: Int,
 
-    val datetime: Int,
+    val datetime: Long,
 
     val amount: String,
 
     val comment: String,
 
-    val id: Int,
+    val id: Long,
 
     val type: Int,
